refactor: cleanup theming constants and types
