+++ frontend/src/metabase/embedding-sdk/theme/MetabaseThemeV2.ts
@@ -1,14 +1,14 @@
 import type {
   ChartColorV2,
-  InternalColorKey,
   MetabaseColorKey,
+  ProtectedColorKey,
 } from "metabase/lib/colors";
 
 /**
  * Colors exposed to Modular Embedding.
  * Excludes colors that are internal, such as admin and Metabase brand colors.
  */
-type MetabaseEmbeddingColorKeyV2 = Exclude<MetabaseColorKey, InternalColorKey>;
+type MetabaseEmbeddingColorKeyV2 = Exclude<MetabaseColorKey, ProtectedColorKey>;
 
 /**
  * V2 theme configuration for embedded Metabase components.
