+++ frontend/src/metabase/lib/colors/index.ts
@@ -3,12 +3,13 @@ export * from "./colors";
 export * from "./theme";
 
 // Re-export from submodules for convenience
-export { INTERNAL_COLORS, type InternalColorKey } from "./constants";
-export { METABASE_DARK_THEME } from "./dark";
-export { METABASE_LIGHT_THEME } from "./light";
+export { PROTECTED_COLORS } from "./constants/protected-colors";
+export { METABASE_DARK_THEME } from "./constants/dark";
+export { METABASE_LIGHT_THEME } from "./constants/light";
 export type {
   ChartColorV2,
   MetabaseColorKey,
   MetabaseThemeV2,
   UserThemeOverride,
+  ProtectedColorKey,
 } from "./types";
